Refactor folder structure to accomodate new diligence component
