+++ src/components/Routes/Fund/FundDiligence/FundDetails.query.ts
