+++ src/components/Routes/Fund/FundOverview/FundOverview.tsx
@@ -1,12 +1,13 @@
 import React from 'react';
 import { FundHoldings } from '../FundHoldings/FundHoldings';
-import { FundPolicies } from '../FundPolicies/FundPolicies';
-import { FundContracts } from './FundContracts/FundContracts';
-import { FundFactSheet } from './FundFactSheet/FundFactSheet';
+import { FundPolicies } from '../FundDiligence/FundPolicies/FundPolicies';
+import { FundContracts } from '../FundDiligence/FundContracts/FundContracts';
+import { FundFactSheet } from '../FundDiligence/FundFactSheet/FundFactSheet';
 import { NewFundPerformanceChart } from '../FundPerformanceChart/FundPerformanceChart';
 import { Grid, GridRow, GridCol } from '~/storybook/Grid/Grid';
 import { FundPerformanceTable } from '~/components/Routes/Fund/FundPerfomanceTable/FundPerformanceTable';
-import { FundPerformanceMetrics } from './FundFactSheet/FundPerformanceMetrics';
+import { FundPerformanceMetrics } from '../FundReturns/FundPerformanceMetrics';
+import { FundDiligence } from '../FundDiligence/FundDiligence';
 
 export interface FundOverviewProps {
   address: string;
@@ -19,19 +20,12 @@ export const FundOverview: React.FC<FundOverviewProps> = ({ address }) => (
         <NewFundPerformanceChart address={address} />
       </GridCol>
     </GridRow>
-    <GridRow>
+    {/* <GridRow>
       <GridCol xs={12} sm={12}>
         <FundPerformanceMetrics address={address} />
       </GridCol>
-    </GridRow>
-    <GridRow>
-      <GridCol xs={12} sm={5}>
-        <FundFactSheet address={address} />
-      </GridCol>
-      <GridCol xs={12} sm={7}>
-        <FundHoldings address={address} />
-      </GridCol>
-    </GridRow>
+    </GridRow> */}
+
     <GridRow>
       <GridCol>
         <FundPerformanceTable address={address} />
@@ -39,10 +33,10 @@ export const FundOverview: React.FC<FundOverviewProps> = ({ address }) => (
     </GridRow>
     <GridRow>
       <GridCol xs={12} sm={6}>
-        <FundPolicies address={address} />
+        <FundHoldings address={address} />
       </GridCol>
       <GridCol xs={12} sm={6}>
-        <FundContracts address={address} />
+        <FundDiligence address={address} />
       </GridCol>
     </GridRow>
   </Grid>
